<refactor> ChessView 인터페이스 추출
