+++ src/test/java/softeer2nd/chess/ChessViewTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import softeer2nd.chess.exceptions.IllegalCommandException;
-import softeer2nd.chess.view.ChessView;
+import softeer2nd.chess.view.ChessConsoleView;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
@@ -15,11 +15,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class ChessViewTest {
 
-    ChessView chessView;
+    ChessConsoleView chessView;
 
     @BeforeEach
     void setup() {
-        chessView = new ChessView();
+        chessView = new ChessConsoleView();
     }
 
     @ParameterizedTest
